% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_json.R
\name{CristinJson}
\alias{CristinJson}
\title{Convert Cristin monthly data to JSON format}
\usage{
CristinJson(
  monthlies,
  unit.paths,
  sdg.data = NULL,
  local.storage = NULL,
  user.cards = TRUE,
  use.multisession = FALSE,
  min.multisession = 25,
  n.workers = NULL,
  n.chunks = NULL,
  handler = "cli",
  restore.defaults = FALSE,
  full.update = FALSE,
  lang = "nn",
  silent = FALSE,
  log = list()
)
}
\arguments{
\item{monthlies}{Data containing monthly information from Cristin.}

\item{unit.paths}{Data containing the Zotero paths of Cristin units.}

\item{sdg.data}{Data containing information about Sustainable Development
Goals (SDGs). Default is NULL.}

\item{local.storage}{Path to local storage directory. Default is NULL.}

\item{user.cards}{Logical, indicating whether to update user information
from INN. Default is TRUE.}

\item{use.multisession}{Logical. If \code{TRUE} (default), parallel
processing using multisession is employed; otherwise, processing is sequential.}

\item{min.multisession}{Minimum number of results for using multisession.
Default: 25}

\item{n.workers}{Optional integer for the number of workers to be used in
multisession mode. If \code{NULL}, it defaults to the number of available
cores minus one (with a minimum of one).}

\item{n.chunks}{Optional integer for the number of chunks to process.
If \code{NULL}, it defaults to the number of workers.}

\item{handler}{The progress handler to be used by the \code{progressr}
package. If \code{NULL} and
  \code{silent} is \code{FALSE}, it defaults to \code{"txtprogressbar"}.
  When \code{silent} is \code{TRUE},
  the handler is set to \code{"void"}.}

\item{restore.defaults}{Logical. If \code{TRUE} (default), the current
\code{future} plan is saved and restored upon exit.}

\item{full.update}{Logical, indicating whether to perform a full update of
user information from INN. Default is FALSE.}

\item{lang}{Language for localization ("nb", "nn", "no", or "en"). Defaults
to "nn".}

\item{silent}{Logical, indicating whether to suppress log messages. Default
is FALSE.}

\item{log}{List to store log messages. Default is an empty list.}
}
\value{
A JSON-formatted string representing the converted Cristin monthly
data.
}
\description{
This function converts the provided Cristin monthly data into
JSON format, which can be used for various purposes.
}
\details{
Used with `CristinMonthly` to create month-to-month bibliography in
JSON format
}
\examples{
\donttest{
  # Create monthlies for unit 209.5.10.0
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "Io1bwAiOjB1jPgUNegjJhJxF",
      user = FALSE
    ),
    unit.id = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = TRUE,
    use.citeproc = FALSE
  ) |>
    c2z:::GoFish()

  # Create json data
  example.json <- CristinJson(
    example$monthlies,
    example$unit.paths,
    user.cards = FALSE
  ) |>
    jsonlite::prettify() |>
    c2z:::GoFish()

  if (any(!is.na(example.json))) cat(example.json)
}
}
