% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_mail.R
\name{CristinMail}
\alias{CristinMail}
\title{Create HTML Email Content from Cristin Data}
\usage{
CristinMail(
  unit.id,
  cristin.monthly,
  archive.host,
  subject = NULL,
  header = NULL,
  footer = NULL,
  replace.style = TRUE,
  width = 700,
  lang = "nn",
  get.ezproxy = FALSE,
  ezproxy.host = "inn.no",
  silent = FALSE
)
}
\arguments{
\item{unit.id}{The key of the unit for which the email content is generated.}

\item{cristin.monthly}{Data containing monthly information from Cristin.}

\item{archive.host}{Host of the archive.}

\item{subject}{The subject of the email. If not provided, it will be
generated based on the data.}

\item{header}{The header of the email. If not provided, it will be generated
based on the data.}

\item{footer}{The footer of the email. If not provided, a default footer
will be used.}

\item{replace.style}{Logical, indicating whether to replace styles in the
generated HTML.}

\item{width}{Width of the email content in pixels.}

\item{lang}{Language for localization ("nb", "nn", "no", or "en"). Defaults
to "nn".}

\item{get.ezproxy}{Use ezproxy, Default: FALSE}

\item{ezproxy.host}{ezproxy host, Default: 'inn.no'}

\item{silent}{c2z is noisy, tell it to be quiet, Default: FALSE}
}
\value{
A list containing the subject and HTML body of the email.
}
\description{
This function generates HTML content for an email using data
from Cristin.
}
\details{
Used with `CristinMonthly` to create month-to-month bibliography in
HTML format for email of selected units
}
\examples{
\donttest{
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "RqlAmlH5l1KPghfCseAq1sQ1",
      user = FALSE
    ),
    unit.id = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = FALSE,
    use.identifiers = FALSE,
    get.unpaywall = FALSE,
    get.ezproxy = FALSE
  )

  # Create HTML email
  if (any(nrow(example$monthlies))) {
    example.mail <- CristinMail("209.5.10.0", example, "example.host")

    # Show subject
    cat(example.mail$subject)
  }
}
}
