% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_monthly.R
\name{CristinMonthly}
\alias{CristinMonthly}
\title{Create a month-to-month bibliography of selected units}
\usage{
CristinMonthly(
  zotero,
  unit.id,
  unit.recursive = TRUE,
  sdg.host = NULL,
  sdg.batch = 20,
  get.unpaywall = FALSE,
  get.ezproxy = FALSE,
  ezproxy.host = "inn.no",
  local.storage = NULL,
  style = "apa-single-spaced",
  locale = "nn-NO",
  start.date = Sys.Date(),
  end.date = NULL,
  use.filter = TRUE,
  filter = NULL,
  nvi = TRUE,
  check.items = TRUE,
  use.identifiers = TRUE,
  use.multisession = FALSE,
  n.workers = NULL,
  n.chunks = NULL,
  handler = NULL,
  restore.defaults = TRUE,
  full.update = FALSE,
  lang = "nn",
  post.lang = "nn",
  post = FALSE,
  post.only = FALSE,
  silent = FALSE,
  cristin.silent = TRUE,
  log = list()
)
}
\arguments{
\item{zotero}{What Zotero library to use}

\item{unit.id}{What unit to search for}

\item{unit.recursive}{Find subunits of defined unit key, Default: TRUE}

\item{sdg.host}{host conducting SDG predictions, Default: NULL}

\item{sdg.batch}{The batch size for each API call to the SDG host, specifying
the number of items to be processed per request., Default: 20}

\item{get.unpaywall}{Find Unpaywall resources, Default: FALSE}

\item{get.ezproxy}{Use ezproxy, Default: FALSE}

\item{ezproxy.host}{ezproxy host, Default: 'inn.no'}

\item{local.storage}{Path to local storage of collections, items and
bibliography, Default: 'NULL'}

\item{style}{Citation style to use for appended bibliography and/or
citations, Default: 'apa-single-spaced'}

\item{locale}{Desired language format of bibliography, Default: 'nn-NO'}

\item{start.date}{Results created from specified date (YYYY-MM),
Default: Sys.Date()}

\item{end.date}{Results created before specified date (YYYY-MM),
Default: NULL}

\item{use.filter}{Filter out specific items (otherwise supported by
`CristinWrangler`), Default: TRUE}

\item{filter}{Filter out specific item types, will default to item types
usually associated with NVI, Default: NULL}

\item{nvi}{Filter out Cristin items not 1/2 in NVI, Default: TRUE}

\item{check.items}{Examine items for duplicates and multidepartemental
publications, Default: TRUE}

\item{use.identifiers}{Use if ISBN/DOI identifiers if enabled, Default: TRUE}

\item{use.multisession}{Logical. If \code{TRUE} (default), parallel
processing using multisession is employed; otherwise, processing is sequential.}

\item{n.workers}{Optional integer for the number of workers to be used in
multisession mode. If \code{NULL}, it defaults to the number of available
cores minus one (with a minimum of one).}

\item{n.chunks}{Optional integer for the number of chunks to process.
If \code{NULL}, it defaults to the number of workers.}

\item{handler}{The progress handler to be used by the \code{progressr}
package. If \code{NULL} and
  \code{silent} is \code{FALSE}, it defaults to \code{"txtprogressbar"}.
  When \code{silent} is \code{TRUE},
  the handler is set to \code{"void"}.}

\item{restore.defaults}{Logical. If \code{TRUE} (default), the current
\code{future} plan is saved and restored upon exit.}

\item{full.update}{Update bibliography for all items in library,
Default: FALSE}

\item{lang}{Define bibliography language (nn, nb, en), Default: 'nn'}

\item{post.lang}{Define language for Zotero collections (nn, nb, en),
Default: 'nn'}

\item{post}{Post new items to specified Zotero library, Default: FALSE}

\item{post.only}{Post new items and nothing more, Default: FALSE}

\item{silent}{c2z is noisy, tell it to be quiet, Default: FALSE}

\item{cristin.silent}{keep queries to Cristin quiet Default: TRUE}

\item{log}{A list for storing log elements, Default: list()}
}
\value{
A list with bibliography and information for defined units
}
\description{
Create a bibliography in a newsletter format from month-to-month
From either a specified unit or a set of units (e.g., A University ->
Faculties -> Departments -> Groups).
}
\details{
Used with `CristinUnits` to create month-to-month bibliography of
selected units
}
\examples{
\donttest{
  # Create monthlies for unit 209.5.10.0
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "Io1bwAiOjB1jPgUNegjJhJxF",
      user = FALSE
    ),
    unit.id = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = TRUE
  )

  # Print the three first individual pages (if any)
  if (!is.null(example$monthlies)) {
    example$monthlies |>
      dplyr::select(title) |>
      print(n = 3, width = 80)
  }
}
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}},
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{bind_rows}},
 \code{\link[dplyr]{bind_cols}}, \code{\link[dplyr]{select}},
 \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{case_when}},
 \code{\link[dplyr]{context}}, \code{\link[dplyr]{distinct}},
 \code{\link[dplyr]{pull}}, \code{\link[dplyr]{mutate-joins}},
 \code{\link[dplyr]{join_by}}, \code{\link[dplyr]{rename}},
 \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}},
 \code{\link[dplyr]{mutate_all}}
 \code{\link[purrr]{pmap}}, \code{\link[purrr]{map}}
 \code{\link[tibble]{tibble}}
 \code{\link[stats]{setNames}}
}
