% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_web.R
\name{CristinWeb}
\alias{CristinWeb}
\title{Create HTML and Metadata Files from Cristin Data for Web Display}
\usage{
CristinWeb(
  monthlies,
  sdg.data = NULL,
  sdg.path = NULL,
  archive.url = NULL,
  local.storage = NULL,
  user.cards = TRUE,
  md.path = "",
  md.save = FALSE,
  use.multisession = FALSE,
  min.multisession = 25,
  n.workers = NULL,
  handler = "cli",
  restore.defaults = FALSE,
  full.update = FALSE,
  lang = "nn",
  silent = FALSE,
  log = list()
)
}
\arguments{
\item{monthlies}{Data containing monthly information from Cristin.}

\item{sdg.data}{Data containing information about Sustainable Development Goals (SDGs). Default is \code{NULL}.}

\item{sdg.path}{Path to SDG images. Default is \code{NULL}.}

\item{archive.url}{Relative URL to the archive. Default is \code{NULL}.}

\item{local.storage}{Path to the local storage directory. Default is \code{NULL}.}

\item{user.cards}{Logical, indicating whether to update user information from INN. Default is \code{TRUE}.}

\item{md.path}{Path to output markdown files. Default is an empty string (current directory).}

\item{md.save}{Logical, indicating whether to save the markdown files to disk. Default is \code{FALSE}.}

\item{use.multisession}{Logical. If \code{TRUE}, parallel processing using multisession is employed;
otherwise, processing is sequential. Default is \code{FALSE}.}

\item{min.multisession}{Minimum number of results for using multisession. Default is \code{25}.}

\item{n.workers}{Optional integer for the number of workers to be used in multisession mode.
If \code{NULL}, it defaults to the number of available cores minus one (with a minimum of one).}

\item{handler}{The progress handler to be used by the \code{progressr} package. Default is \code{"cli"}.}

\item{restore.defaults}{Logical. If \code{TRUE}, the current \code{future} plan is saved and restored upon exit.
Default is \code{FALSE}.}

\item{full.update}{Logical, indicating whether to perform a full update of user information from INN.
Default is \code{FALSE}.}

\item{lang}{Language for localization ("nb", "nn", "no", or "en"). Defaults to \code{"nn"}.
If \code{lang} is set to \code{"no"}, it is internally converted to \code{"nn"}. Any language not in
\code{c("nb", "nn", "no")} is set to \code{"en"}.}

\item{silent}{Logical, indicating whether to suppress log messages. Default is \code{FALSE}.}

\item{log}{List to store log messages. Default is an empty list.}
}
\value{
A list containing two elements:
\item{results}{A tibble with metadata and HTML content (as generated by \code{htmltools}) for each publication.}
\item{log}{A log of processing messages.}
}
\description{
This function processes monthly Cristin data to generate HTML content
and metadata files for web display. It uses \code{htmltools} to construct interactive HTML
elements (e.g., buttons, articles) for each publication, and optionally writes markdown files
with frontmatter for further processing.
}
\details{
This function is typically used with \code{CristinMonthly} to convert monthly publication data
into HTML components and markdown files with frontmatter metadata. The function leverages \code{htmltools}
to programmatically build HTML elements such as articles and buttons for links to additional publication information.
}
\examples{
\donttest{
  # Create monthlies for unit 209.5.10.0
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "Io1bwAiOjB1jPgUNegjJhJxF",
      user = FALSE
    ),
    unit.id = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = TRUE,
    use.citeproc = FALSE
  )

  # Generate HTML and markdown files
  example.web <- CristinWeb(
    example$monthlies,
    user.cards = FALSE
  )

  # Print the three first publication titles (if any)
  if (!is.null(example$monthlies)) {
    example$monthlies |>
      dplyr::select(title) |>
      print(n = 3, width = 80)
  }
}
}
