% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_web.R
\name{CristinWeb}
\alias{CristinWeb}
\title{Create HTML and Metadata Files from Cristin Data for Web Display}
\usage{
CristinWeb(
  monthlies,
  sdg.data = NULL,
  sdg.path = NULL,
  archive.url = NULL,
  local.storage = NULL,
  user.cards = TRUE,
  full.update = FALSE,
  lang = "nn",
  silent = FALSE,
  log = list()
)
}
\arguments{
\item{monthlies}{Data containing monthly information from Cristin.}

\item{sdg.data}{Data containing information about Sustainable Development
Goals (SDGs). Default is NULL.}

\item{sdg.path}{Path to SDG images. Default is NULL.}

\item{archive.url}{Relative URL to archive. Default is NULL.}

\item{local.storage}{Path to local storage directory. Default is NULL.}

\item{user.cards}{Logical, indicating whether to update user information
from INN. Default is TRUE.}

\item{full.update}{Logical, indicating whether to perform a full update of
user information from INN. Default is FALSE.}

\item{lang}{Language for localization ("nb", "nn", "no", or "en"). Defaults
to "nn".}

\item{silent}{Logical, indicating whether to suppress log messages. Default
is FALSE.}

\item{log}{List to store log messages. Default is an empty list.}
}
\value{
A tibble containing metadata and HTML content for web display.
}
\description{
This function generates HTML and metadata files for web
display using data from Cristin.
}
\details{
Used with `CristinMonthly` to create month-to-month bibliography in
markdown format
}
\examples{
\donttest{
  # Create monthlies for unit 209.5.10.0
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "Io1bwAiOjB1jPgUNegjJhJxF",
      user = FALSE
    ),
    unit.id = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = TRUE
  )

  # Create md files
  example.web <- CristinWeb(
    example$monthlies,
    user.cards = FALSE
  )

  # Print the three first individual pages (if any)
  if (!is.null(example$monthlies)) {
    example$monthlies |>
      dplyr::select(title) |>
      print(n = 3, width = 80)
  }
}
}
