% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_web.R
\name{CristinWeb}
\alias{CristinWeb}
\title{Create HTML and Metadata Files from Cristin Data for Web Display}
\usage{
CristinWeb(
  monthlies,
  sdg.data = NULL,
  sdg.path = NULL,
  archive.url = NULL,
  local.storage = NULL,
  user.cards = TRUE,
  use.multisession = FALSE,
  min.multisession = 25,
  n.workers = NULL,
  handler = "cli",
  restore.defaults = FALSE,
  full.update = FALSE,
  lang = "nn",
  silent = FALSE,
  log = list()
)
}
\arguments{
\item{monthlies}{Data containing monthly information from Cristin.}

\item{sdg.data}{Data containing information about Sustainable Development
Goals (SDGs). Default is NULL.}

\item{sdg.path}{Path to SDG images. Default is NULL.}

\item{archive.url}{Relative URL to archive. Default is NULL.}

\item{local.storage}{Path to local storage directory. Default is NULL.}

\item{user.cards}{Logical, indicating whether to update user information
from INN. Default is TRUE.}

\item{use.multisession}{Logical. If \code{TRUE} (default), parallel
processing using multisession is employed; otherwise, processing is sequential.}

\item{min.multisession}{Minimum number of results for using multisession.
Default: 25}

\item{n.workers}{Optional integer for the number of workers to be used in
multisession mode. If \code{NULL}, it defaults to the number of available
cores minus one (with a minimum of one).}

\item{handler}{The progress handler to be used by the \code{progressr}
package. If \code{NULL} and
  \code{silent} is \code{FALSE}, it defaults to \code{"txtprogressbar"}.
  When \code{silent} is \code{TRUE},
  the handler is set to \code{"void"}.}

\item{restore.defaults}{Logical. If \code{TRUE} (default), the current
\code{future} plan is saved and restored upon exit.}

\item{full.update}{Logical, indicating whether to perform a full update of
user information from INN. Default is FALSE.}

\item{lang}{Language for localization ("nb", "nn", "no", or "en"). Defaults
to "nn".}

\item{silent}{Logical, indicating whether to suppress log messages. Default
is FALSE.}

\item{log}{List to store log messages. Default is an empty list.}
}
\value{
A list containing metadata and HTML content for web display.
}
\description{
This function generates HTML and metadata files for web
display using data from Cristin.
}
\details{
Used with `CristinMonthly` to create month-to-month bibliography in
markdown format
}
\examples{
\donttest{
  # Create monthlies for unit 209.5.10.0
  example <- CristinMonthly(
    c2z::Zotero(
      id = "4827927",
      api = "Io1bwAiOjB1jPgUNegjJhJxF",
      user = FALSE
    ),
    unit.id = "209.5.10.0",
    start.date = "2023-07",
    post = TRUE,
    silent = TRUE,
    use.citeproc = FALSE,
  )

  # Create md files
  example.web <- CristinWeb(
    example$monthlies,
    user.cards = FALSE
  )

  # Print the three first individual pages (if any)
  if (!is.null(example$monthlies)) {
    example$monthlies |>
      dplyr::select(title) |>
      print(n = 3, width = 80)
  }
}
}
